backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库和存储
supabase==2.0.2
redis==5.0.1

# AI模型
google-generativeai==0.3.2

# 数据处理和验证
pydantic==2.5.0
pydantic-settings==2.1.0

# 安全和身份认证
PyJWT==2.8.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4

# 日志和监控
loguru==0.7.2
prometheus-client==0.19.0

# 文件处理
openpyxl==3.1.2
python-multipart==0.0.6

# 工具库
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 类型检查
mypy==1.7.1

# 代码格式化
black==23.11.0
isort==5.12.0
flake8==6.1.0